chore: removing junk, install packages, tidy folder structure
