+++ app/layout.tsx
@@ -1,0 +1,34 @@
+import type { Metadata } from "next";
+// UTILS
+import { roboto, poppins } from "@/utils/fonts";
+import { cn } from "@/utils/helpers";
+// STYLES
+import '@/styles/global.css';
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <head>
+      </head>
+      <body 
+        className={cn(
+          "min-h-screen bg-secondary font-roboto antialiased",
+          roboto.variable,
+          poppins.variable
+        
+        )}
+      >
+    {children}
+      </body>
+    </html>
+  );
+}
