+++ app/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <main className="container  px-5 pt-5">
+      <h1>Welcome to Dashboard page</h1>
+    </main>
+  );
+}
