--- components/SVGs/drill-and-wrench.tsx
@@ -1,18 +1,0 @@
-// REACT
-import * as React from "react"
-// TYPES
-import { SVGPropType } from "@/types";
-
-
-export const DrillAndWrenchSVG = ({ width, height, fill}: SVGPropType) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    xmlSpace="preserve"
-    width={width}
-    height={height}
-  >
-    <g fill={fill}>
-      <path d="m15.71 14.931-2.387-7.787-8.032-1.868-1.193 1.273.022.019.007-.007.175.161.001.001 5.049 4.642v2.564l-2.822.503-5.099-4.69-.202-.133L.02 10.898l2.464 7.854 7.908 1.899 21.99 22.073 5.453-5.434L16.35 15.724zM6.412 37.29l5.455 5.432 8.216-8.241-5.435-5.453zM47.98 18.694l-3.422-3.483-.989.972-4.086-4.101-.009-.008c-8.008-8.196-16.667-6.601-16.667-6.601l7.66 7.688-6.283 6.302 5.434 5.453 6.283-6.302 2.569 2.58-.975.958 3.421 3.482.999-.981.065.064 5.064-5.045-.031-.03.967-.948z" />
-    </g>
-  </svg>
-)
