--- components/SVGs/index.tsx
@@ -1,18 +1,0 @@
-// COMPONENTS
-import { PaintBrushSVG } from "./paint-brush";
-import { DrillAndWrenchSVG } from "./drill-and-wrench";
-import { TillingSVG } from "./tilling";
-import { MinusSymbolSVG } from "./minus";
-import { PlusSymbolSVG } from "./plus";
-import { InstagramSVG } from "./instagram";
-import { FacebookSVG } from "./facebook";
-
-export { 
-  PaintBrushSVG, 
-  DrillAndWrenchSVG, 
-  TillingSVG, 
-  MinusSymbolSVG,
-  PlusSymbolSVG,
-  InstagramSVG,
-  FacebookSVG
-};
