--- config/faq.ts
@@ -1,53 +1,0 @@
-export const faq = [
-  {
-    id: 1,
-    question: 'How much do you charge for your services?',
-    answer: 'Our pricing varies depending on the scope of the project, materials required, and other factors. We offer free estimates, so feel free to contact us for a personalized quote.'
-  },
-  {
-    id: 2,
-    question: 'What areas do you serve?',
-    answer: `We provide our services to clients in Newcastle, Gateshead, Sunderland, and subsequent regions. If you're unsure whether we cover your location, please reach out to us, and we'll be happy to assist you.`
-  },
-  {
-    id: 3,
-    question: 'Do you provide materials, or do I need to purchase them separately?',
-    answer: 'We can provide materials for your project, or you can choose to purchase them yourself. Our team is experienced in working with various materials and can offer guidance on what would work best for your needs.'
-  },
-  {
-    id: 4,
-    question: 'How long will it take to complete my project?',
-    answer: `The timeframe for completion depends on the size and complexity of the project. We strive to complete projects efficiently without compromising on quality. Once we assess your requirements, we'll provide you with a timeline for completion.`
-  },
-  {
-    id: 5,
-    question: 'How do I schedule an appointment or request a quote?',
-    answer: `You can easily schedule an appointment or request a quote by filling out the contact form on our website. Alternatively, you can give us a call or send us an email, and we'll get back to you promptly.`
-  },
-  // Reminder to add these back once we have the services
-  // {
-  //   id: 6,
-  //   question: 'Do you offer any warranties or guarantees on your work?',
-  //   answer: `Yes, we stand behind the quality of our workmanship. We offer warranties on our services to ensure your satisfaction. If you encounter any issues after the completion of the project, please don't hesitate to reach out to us, and we'll make it right.`
-  // },
-  // {
-  //   id: 7,
-  //   question: 'Are you licensed and insured?',
-  //   answer: `Yes, we are fully licensed and insured to provide our services. You can have peace of mind knowing that we adhere to industry standards and regulations, and your property is protected throughout the project.`
-  // },
-  {
-    id: 8,
-    question: 'Can you provide references or examples of your previous work?',
-    answer: `Absolutely! We have a portfolio of our past projects that showcase the quality of our work. We can also provide references from satisfied clients upon request.`
-  },
-  {
-    id: 9,
-    question: 'Do I need to prepare my space before you begin work?',
-    answer: `We recommend clearing the area of any furniture or belongings to facilitate the work process. If you need assistance with this, please let us know, and we can make arrangements to help.`
-  },
-  {
-    id: 10,
-    question: 'What payment methods do you accept?',
-    answer: `We accept various payment methods for your convenience, including cash, credit/debit cards, and bank transfers. Payment details will be discussed and agreed upon before the start of the project.`
-  },
-];
