--- config/selling-points.ts
@@ -1,17 +1,0 @@
-export const sellingPoints = [
-  {
-    id: 1,
-    title: 'Painting',
-    description: `Transform your space with precision and expertise. From refreshing walls with vibrant hues to intricate detailing, our professional painting services breathe new life into any room. With a keen eye for color and quality, we ensure a flawless finish that enhances the aesthetic appeal of your home.`
-  },
-  {
-    id: 2,
-    title: 'Tiling',
-    description: `Elevate your kitchen, bathroom, or flooring with expert tiling craftsmanship. Our meticulous attention to detail and skilled techniques result in stunning tiled surfaces that marry durability with beauty. Whether it's intricate mosaic patterns or sleek modern designs, our tiling solutions bring elegance and sophistication to your space.`
-  },
-  {
-    id: 3,
-    title: 'Miscelleneous',
-    description: `Enhance your outdoor spaces with our versatile services. From sturdy brickwork to impeccably landscaped gardens, we offer comprehensive solutions tailored to your needs. Whether it's reviving your driveway, creating an oasis of greenery, or securing your property with durable fencing, our expertise ensures impeccable results that stand the test of time.`
-  },
-];
