--- content/projects/bathroom-tiling.mdx
@@ -1,52 +1,0 @@
----
-title: Luxury Redefined - Bathroom Tiling Project
-description: Total bathroom remodelling with tiles covering walls and floor. Project included moisture control prior
-categories:
-  - 
-    variant: tiling
-thumbnailImg: /images/projects/bathroom-tiling/after.jpeg
-thumbnailImgAlt: bathroom tiling project thumbnail
-date: "2023-01-08"
----
-
-**Client:** Mr. and Mrs Hoxha
-
-**Location:** "Lobely Hill, Gateshead
-
-**Scope of Work:**
-- **Area Covered:** Entire bathroom including walls and floor
-- **Steps:**
-  1. Initial Consultation: Discussed the client's vision and preferences for tile material, color, and design.
-  2. Measurement and Planning: Accurate measurements were taken to ensure precise installation and minimize waste.
-  3. Surface Preparation: Properly cleaned and leveled the surfaces to ensure a smooth and even tiling foundation.
-  4. Tile Selection: Assisted the client in selecting high-quality tiles that complemented the overall aesthetic of the bathroom.
-  5. Installation: Meticulously installed the tiles with precision and expertise, paying close attention to patterns and grout lines.
-  6. Finishing Touches: Applied sealant to protect the tiles and enhance their longevity and appearance.
-
-**Outcome:**
-- The bathroom walls are adorned with sleek, modern tiles, creating a sophisticated ambiance.
-- The floor features durable and elegant tiles that add texture and visual interest to the space.
-- The seamless integration of tiles enhances the sense of spaciousness and luxury in the bathroom.
-- The color scheme and pattern selection reflect the client's refined taste and elevate the overall aesthetic appeal of the room.
-
-**Before:**
-
-<Image
-  src="/images/projects/bathroom-tiling/before.jpeg"
-  width="718"
-  height="404"
-  alt="Image"
-/>
-
-**After:**
-
-<Image
-  src="/images/projects/bathroom-tiling/after.jpeg"
-  width="718"
-  height="404"
-  alt="Image"
-/>
-
-**Customer Testimonial:**
-
-*"I couldn't be happier with the incredible transformation of my bathroom by [Your Company Name] Tiling Solutions. From start to finish, they were professional, efficient, and attentive to detail. The result is a stunning bathroom that exceeds my expectations. Highly recommend!"* - Ms. Emily Roberts
