--- content/projects/driveway-transformation.mdx
@@ -1,53 +1,0 @@
----
-title: Contemporary Driveway Transformation
-description: Converting a garden section into a modern driveway with brick runway and exterior brickwall.
-categories:
-  - 
-    variant: landscaping
-thumbnailImg: /images/projects/driveway-transformation/after.png
-thumbnailImgAlt: bathroom tiling project thumbnail
-date: "2023-05-15"
----
-
-**Client:** Mr. Panagiotis
-
-**Location:** Lobley Hill, Gateshead
-
-**Scope of Work:**
-- **Area Covered:** Existing garden section transformed into a functional driveway.
-- **Steps:**
-  1. Initial Consultation: Discussed client's requirements and assessed the space for driveway feasibility.
-  2. Design Proposal: Presented modern design concepts integrating brick runway and exterior brickwall.
-  3. Material Selection: Assisted the client in selecting high-quality bricks suitable for driveway construction.
-  4. Ground Preparation: Cleared and leveled the area to ensure a solid foundation for the driveway.
-  5. Runway Installation: Laid bricks in a patterned runway design, ensuring durability and aesthetic appeal.
-  6. Brickwall Construction: Erected an exterior brickwall to enhance the driveway's boundary and aesthetics.
-  7. Finishing Touches: Applied sealant to the bricks for protection against weathering and to maintain longevity.
-
-**Outcome:**
-- The driveway showcases a contemporary design that seamlessly integrates with the surrounding landscape.
-- The brick runway adds visual interest and texture, enhancing the overall curb appeal of the property.
-- The exterior brickwall provides privacy and security while complementing the architectural style of the house.
-- The driveway transformation maximizes functionality and adds value to the property.
-
-**Before:**
-
-<Image
-  src="/images/projects/driveway-transformation/before.png"
-  width="718"
-  height="404"
-  alt="Image"
-/>
-
-**After:**
-
-<Image
-  src="/images/projects/driveway-transformation/after.png"
-  width="718"
-  height="404"
-  alt="Image"
-/>
-
-**Customer Testimonial:**
-
-*"They transformed our garden into a stunning driveway that perfectly suits our needs and enhances the aesthetics of our home. Professionalism and attention to detail were evident throughout the project. Highly recommend!"* - Panagiotis Mita
