+++ package.json
@@ -19,7 +19,6 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.50.1",
-    "rehype-slug": "^6.0.0",
     "tailwind-merge": "^2.2.1"
   },
   "devDependencies": {
