--- styles/contact-page.css
@@ -1,15 +1,0 @@
-.backdrop-banner{
-  flex: 1;
-}
-
-@media screen and (max-width: 768px) {
-  .backdrop-banner{
-    position: absolute;
-    width: 100%;
-    height: 90vh;
-    left: 0;
-    top: 10vh;
-  }
-  
-}
-
